refactor : 레코드로 변경
