+++ src/main/java/gift/Application.java
@@ -8,7 +8,7 @@ import org.springframework.retry.annotation.EnableRetry;
 
 @EnableRetry
 @EnableJpaAuditing
-@ConfigurationPropertiesScan("gift.config")
+@ConfigurationPropertiesScan
 @SpringBootApplication
 public class Application {
     public static void main(String[] args) {
