+++ src/main/resources/application.properties
@@ -9,3 +9,8 @@ spring.jpa.hibernate.ddl-auto=create
 jwt.secretKey=secretKey
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
+kakao.rest_api_key=a9e83e916ad54f35fba73a89103c854d
+kakao.redirect_uri=http://localhost:8080/kakao/auth
+kakao.token_url=https://kauth.kakao.com/oauth/token
+kakao.auth_url=https://kauth.kakao.com/oauth/authorize
+kakao.user_info_url=https://kapi.kakao.com/v2/user/me
